added local version check

+++ updateplex.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
 #This script will log into plex and download the newest plexpass version for Ubnutu 64 bit
 if [ ! -f variables.php ]; then
-    echo "<?php" > variables.php
-    echo "    \$login = '';" >> variables.php
-    echo "    \$password = '';" >> variables.php
-    echo "?>" >> variables.php
-    echo "Please open variables.php in an editor of choice and complete the required fields before running again."
-    exit 1
+    echo "<?php" > variables.php;
+    echo "    \$login = '';" >> variables.php;
+    echo "    \$password = '';" >> variables.php;
+    echo "?>" >> variables.php;
+    echo "Please open variables.php in an editor of choice and complete the required fields before running again.";
+    exit 1;
 fi
 # creating the cookie.txt file for the session
 touch cookie.txt
@@ -14,15 +14,25 @@ chmod 777 cookie.txt
 
 echo "Identifying newest plex version"
 # Running the php command to get the newest plexpass version of plex using the login info in variables.php and setting that path to $URL
-URL=$(php curl_plex.php | grep '="Ubuntu">64-bit</a>' | cut -d'"' -f2)
+webversion=$(php curl_plex.php | grep '="Ubuntu">64-bit</a>' | cut -d'"' -f2)
+currentversion=$(curl -s http://localhost:32400 | grep -Po '(?<=" version=")([^">]*)')
+
+if [[ $webversion == *$currentversion* ]]; then
+    echo "Plex media server is up-to-date!";
+    echo "Current Version is: "$currentversion;
+    exit
+fi
 
 echo "Downloading newest plex verison"
 # Removing the session cookie, cookie.txt
 rm cookie.txt
 # Downloading the file to the users home directory
-wget $URL -P ~/
+wget -N $webversion -P ~/
 
 # Asking the user if they want to install the file or not. If yes then after it will remove the deb file.
+if [$1 == "--force"]; then
+   sudo dpkg -i ~/plexmediaserver*;rm ~/plexmediaserver* ;
+else
 while true; do
     read -p "Do you wish to install the newest version of plex (plexpass)?" yn
     case $yn in
@@ -31,3 +41,4 @@ while true; do
         * ) echo "Please answer yes or no.";;
     esac
 done
+fi
